django/mainApp/requirements.txt
Django==3.1.8
djangorestframework==3.12.2
django_environ==0.11.2
django-cors-headers==3.10.1
psycopg2==2.9.8